# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Laravel OpenApi Demo Documentation'
  description: 'L5 Swagger OpenApi description'
  contact:
    email: admin@admin.com
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8001/api'
    description: 'Demo API Server'
paths:
  /posts:
    get:
      tags:
        - Posts
      summary: 'Get list of posts'
      description: 'Returns list of posts'
      operationId: getPostList
      responses:
        200:
          description: 'Successful operation'
        401:
          description: Unauthenticated
        403:
          description: Forbidden
    post:
      tags:
        - Posts
      summary: 'Store new post'
      description: 'Returns post data'
      operationId: storePost
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              required:
                - image
                - title
                - content
              properties:
                image:
                  description: 'File (jpeg,png,jpg,gif,svg)'
                  type: string
                  format: file
                title:
                  description: Title
                  type: string
                content:
                  description: Content
                  type: string
              type: object
      responses:
        201:
          description: 'Successful operation'
        400:
          description: 'Bad Request'
        401:
          description: Unauthenticated
        403:
          description: Forbidden
  '/posts/{post}':
    get:
      tags:
        - Posts
      summary: 'Get post information'
      description: 'Returns post data'
      operationId: getPostById
      parameters:
        -
          name: post
          in: query
          description: 'Post id'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
        400:
          description: 'Bad Request'
        401:
          description: Unauthenticated
        422:
          description: 'Unprocessable Entity'
    put:
      tags:
        - Posts
      summary: 'Update existing post'
      description: 'Returns updated post data'
      operationId: updatePost
      parameters:
        -
          name: post
          in: query
          description: 'Post id'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              required:
                - image
                - title
                - content
              properties:
                image:
                  description: 'File (jpeg,png,jpg,gif,svg)'
                  type: string
                  format: file
                title:
                  description: Title
                  type: string
                content:
                  description: Content
                  type: string
              type: object
      responses:
        201:
          description: 'Successful operation'
        400:
          description: 'Bad Request'
        401:
          description: Unauthenticated
        422:
          description: 'Unprocessable Entity'
    delete:
      tags:
        - Posts
      summary: 'Delete existing post'
      description: 'Deletes a record and returns no content'
      operationId: deletePost
      parameters:
        -
          name: post
          in: path
          description: 'Post id'
          required: true
          schema:
            type: integer
      responses:
        204:
          description: 'Successful operation'
          content:
            application/json:
              schema: []
        401:
          description: Unauthenticated
        403:
          description: Forbidden
        404:
          description: 'Resource Not Found'
tags:
  -
    name: Posts
    description: 'API Endpoints of Posts'